+++ packages/api-client/src/v2/features/collection/components/Settings.vue
@@ -1,19 +1,12 @@
 <script setup lang="ts">
 import type { ThemeId } from '@scalar/themes'
 import type { ColorMode } from '@scalar/workspace-store/schemas/workspace'
-import { useRouter } from 'vue-router'
 
 import type { CollectionProps } from '@/v2/features/app/helpers/routes'
 import { CollectionSettings, DocumentSettings } from '@/v2/features/settings'
 
-const {
-  eventBus,
-  documentSlug,
-  document,
-  activeWorkspace,
-  workspaceStore,
-  collectionType,
-} = defineProps<CollectionProps>()
+const { eventBus, documentSlug, document, workspaceStore, collectionType } =
+  defineProps<CollectionProps>()
 
 const handleUpdateWatchMode = (watchMode: boolean) => {
   eventBus.emit('document:update:watch-mode', watchMode)
@@ -29,19 +22,6 @@ const handleUpdateActiveProxy = (proxy: string | null) => {
 const handleUpdateColorMode = (colorMode: ColorMode) => {
   eventBus.emit('workspace:update:color-mode', colorMode)
 }
-
-const router = useRouter()
-
-const handleDeleteDocument = () => {
-  workspaceStore.deleteDocument(documentSlug)
-
-  router.push({
-    name: 'workspace.environment',
-    params: {
-      workspaceSlug: activeWorkspace.id,
-    },
-  })
-}
 </script>
 <template>
   <DocumentSettings
@@ -49,8 +29,10 @@ const handleDeleteDocument = () => {
     :documentUrl="document?.['x-scalar-original-source-url']"
     :isDraftDocument="documentSlug === 'drafts'"
     :title="document?.info.title ?? ''"
-    :watchMode="document?.['x-scalar-watch-mode'] ?? true"
-    @delete:document="handleDeleteDocument"
+    :watchMode="document?.['x-scalar-watch-mode']"
+    @delete:document="
+      eventBus.emit('document:delete:document', { name: documentSlug })
+    "
     @update:watchMode="handleUpdateWatchMode" />
   <CollectionSettings
     v-else