+++ packages/api-client/src/v2/features/command-palette/index.ts
@@ -1,0 +1,3 @@
+export { importDocumentToWorkspace } from './helpers/import-document-to-workspace'
+export { loadDocumentFromSource } from './helpers/load-document-from-source'
+export { default as TheCommandPalette } from './TheCommandPalette.vue'