+++ packages/api-reference-react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @scalar/api-reference-react
 
+## 0.8.51
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.44.14**
+  - [#8025](https://github.com/scalar/scalar/pull/8025): fix: use flex-start instead of start for better browser support
+  - [#8056](https://github.com/scalar/scalar/pull/8056): fix: hideModels not applying correctly
+
 ## 0.8.50
 
 ### Patch Changes